+++ crates/bili_sync/src/workflow.rs
@@ -690,76 +690,3 @@ async fn generate_nfo(nfo: NFO<'_>, nfo_path: PathBuf) -> Result<()> {
     fs::write(nfo_path, nfo.generate_nfo().await?.as_bytes()).await?;
     Ok(())
 }
-
-#[cfg(test)]
-mod tests {
-    use handlebars::handlebars_helper;
-    use serde_json::json;
-
-    use super::*;
-
-    #[test]
-    fn test_template_usage() {
-        let mut template = handlebars::Handlebars::new();
-        handlebars_helper!(truncate: |s: String, len: usize| {
-            if s.chars().count() > len {
-                s.chars().take(len).collect::<String>()
-            } else {
-                s.to_string()
-            }
-        });
-        template.register_helper("truncate", Box::new(truncate));
-        let _ = template.path_safe_register("video", "test{{bvid}}test");
-        let _ = template.path_safe_register("test_truncate", "哈哈，{{ truncate title 30 }}");
-        let _ = template.path_safe_register("test_path_unix", "{{ truncate title 7 }}/test/a");
-        let _ = template.path_safe_register("test_path_windows", r"{{ truncate title 7 }}\\test\\a");
-        #[cfg(not(windows))]
-        {
-            assert_eq!(
-                template
-                    .path_safe_render("test_path_unix", &json!({"title": "关注/永雏塔菲喵"}))
-                    .unwrap(),
-                "关注_永雏塔菲/test/a"
-            );
-            assert_eq!(
-                template
-                    .path_safe_render("test_path_windows", &json!({"title": "关注/永雏塔菲喵"}))
-                    .unwrap(),
-                "关注_永雏塔菲_test_a"
-            );
-        }
-        #[cfg(windows)]
-        {
-            assert_eq!(
-                template_borrowed
-                    .path_safe_render("test_path_unix", &json!({"title": "关注/永雏塔菲喵"}))
-                    .unwrap(),
-                "关注_永雏塔菲_test_a"
-            );
-            assert_eq!(
-                template_borrowed
-                    .path_safe_render("test_path_windows", &json!({"title": "关注/永雏塔菲喵"}))
-                    .unwrap(),
-                r"关注_永雏塔菲\\test\\a"
-            );
-        }
-        assert_eq!(
-            template
-                .path_safe_render("video", &json!({"bvid": "BV1b5411h7g7"}))
-                .unwrap(),
-            "testBV1b5411h7g7test"
-        );
-        assert_eq!(
-            template
-                .path_safe_render(
-                    "test_truncate",
-                    &json!({"title": "你说得对，但是 Rust 是由 Mozilla 自主研发的一款全新的编译期格斗游戏。\
-                    编译将发生在一个被称作「Cargo」的构建系统中。在这里，被引用的指针将被授予「生命周期」之力，导引对象安全。\
-                    你将扮演一位名为「Rustacean」的神秘角色, 在与「Rustc」的搏斗中邂逅各种骨骼惊奇的傲娇报错。\
-                    征服她们、通过编译同时，逐步发掘「C++」程序崩溃的真相。"})
-                )
-                .unwrap(),
-            "哈哈，你说得对，但是 Rust 是由 Mozilla 自主研发的一"
-        );
-    }
-}
